refatorando userPost
