+++ src/components/shared/UserPost.js
@@ -1,21 +1,21 @@
-import { Link, useLocation } from 'react-router-dom' 
+import { Link } from 'react-router-dom' 
 import { FiTrash } from "react-icons/fi";
-import { AiOutlineHeart, AiFillHeart } from 'react-icons/ai'
-import ContainerUserPost from './ContainerUserPost'
-import ReactModal from 'react-modal';
-import { deletePost, getPosts, getUserPosts } from '../../service/API';
-import styled from 'styled-components';
+import { AiOutlineHeart, AiFillHeart, AiOutlineComment } from 'react-icons/ai'
+import {ContainerUserPost, HashtagCSS, Interaction, EditBox } from './ContainerUserPost'
 import { toggleLike, editPost } from '../../service/API';
 import UserContext from '../../contexts/UserContext';
 import ReactTooltip from 'react-tooltip';
 import { TiPencil } from 'react-icons/ti';
+import { deletePost, getPosts, getUserPosts } from '../../service/API';
 import { useEffect, useContext, useRef, useState } from 'react';
+import { useLocation } from 'react-router';
+import Repost from './Repost'
+import Modal from './Modal'
 import Swal from 'sweetalert2';
 
-export default function UserPost(props) {
-    let location = useLocation();
-    
 
+export default function UserPost(props) {
+    const location = useLocation();
     const {
         id,
         linkTitle, 
@@ -24,44 +24,23 @@ export default function UserPost(props) {
         linkDescription, 
         link, 
         likes,
+        setPosts
     } = props.post;
-  
-    const { userInfo, userId, setPosts } = props;
+    
     const { user } = useContext(UserContext);
-  
+    const [enabled, setEnabled] = useState(true);
     const [tooltipMessage, setTooltipMessage] = useState('')
     const [liked, setLiked] = useState(likes.some(like => like.userId === user.user.id));
     const [postLikes, setPostLikes] = useState(likes);
-
     const [myPost, setMyPost] = useState(false);
     const [editMode, setEditMode] = useState(false);
     const [actualText, setActualText] = useState(text);
     const [editedText, setEditedText] = useState(text);
     const [isDisabled, setIsDisabled] = useState(false);
-
+    const [isOpen, setIsOpen] = useState(false); 
+    const { userInfo, userId } = props;
     const textAreaRef = useRef();
     const effectTooltip = renderTooltip;
-  
-    const [habilitar,setHabilitar] = useState(true);
-    ReactModal.setAppElement(document.getElementById('root'))
-    const [isOpen,setIsopen] = useState(false);
-    const customStyles = {
-        content: {
-          top: '50%',
-          left: '50%',
-          right: 'auto',
-          bottom: 'auto',
-          marginRight: '-50%',
-          transform: 'translate(-50%, -50%)',
-          background: '#333333',
-          borderRadius: '50px',
-          width: '600px',
-          height: '262px',
-          display:'flex',
-          justifyContent: 'center',
-        },
-      };
-  
     useEffect(() => {
         if(user.user.id === userInfo.id){
                 setMyPost(true);
@@ -81,43 +60,6 @@ export default function UserPost(props) {
             return false;
         }
     }
-       
-    function AbrirModal(){
-        setIsopen(true);
-    }
-
-    function FecharModal(){
-        setIsopen(false);
-    }
-    function ApagarPost(id){
-        setHabilitar(false);
-        console.log(id);
-        deletePost(user.token, id).then(Sucesso).catch(Erro);
-    }
-
-    function Sucesso(){
-        setHabilitar(true);
-        setIsopen(false);
-        alert("Post deletado com sucesso");
-        if(location.pathname === "/timeline"){
-            getPosts(user.token).then((res)=> {
-                setPosts(res.data);
-            })
-        }
-        else if (location.pathname === "/my-posts"){
-            getUserPosts({ token: user.token, userId: user.user.id})
-            .then((r) => {
-                setPosts(r.data);
-            })
-        }
-
-    }
-    function Erro(){
-        setHabilitar(true)
-        setIsopen(false);
-        alert("Não foi possível excluir o Post tenta novamente");
-    }
-
     function checkHashtag() {
         const textCheck = actualText.split(' ').map((word, index) => {
             if (word[0] === '#') {
@@ -212,6 +154,24 @@ export default function UserPost(props) {
         }
     }
 
+    function success(){
+        setEnabled(true);
+        setIsOpen(false);
+        alert("Post deletado com sucesso");
+        if(location.pathname === "/timeline"){
+            getPosts(user.token).then((res)=> {
+                setPosts(res.data);
+            })
+        }
+        else if (location.pathname === "/my-posts"){
+            getUserPosts({ token: user.token, userId: user.user.id})
+            .then((res) => {
+                setPosts(res.data);
+            })
+        }
+
+    }
+
     function checkEditMode(){
         if(editMode){
             setEditedText(actualText)
@@ -253,26 +213,13 @@ export default function UserPost(props) {
 
     return (
         <ContainerUserPost id="main">
-            <ReactModal
-                isOpen={isOpen}
-                onRequestClose={FecharModal}
-                style={customStyles}
-                contentLabel="Example Modal"
-            >
-                    <BoxModal>
-                        <ModalTitle>{habilitar ? 'Tem certeza que deseja excluir essa publicação' : 'Carregando...' }</ModalTitle>
-                        <div>
-                            <ModalCancel onClick={FecharModal} state={habilitar}>Não, voltar</ModalCancel>
-                            <ModalConfirm onClick={()=>ApagarPost(id)}  state={habilitar}>Sim, excluir</ModalConfirm>
-                        </div>
-                    </BoxModal>
-            </ReactModal>
+            <Modal  isOpen={isOpen}  user={user} id={id}  setIsOpen={setIsOpen} success={success} enabled={enabled} setEnabled={setEnabled}/>
             <div className="photo-and-likes">
                 <Link
 					to={`/user/${userId}`}>
 				<img src={userInfo.avatar} alt=''/>
 				</Link>
-                    <Likes data-tip={tooltipMessage} >
+                    <Interaction data-tip={tooltipMessage} >
                         {liked ? 
                             <AiFillHeart
                             style={{color: '#ac0000'}} 
@@ -290,14 +237,22 @@ export default function UserPost(props) {
                             backgroundColor="#FFFFFF30"
                             textColor="#505050"
                         />
-                    </Likes>
+                    </Interaction>
+                    <Interaction  >
+
+                    <AiOutlineComment 
+                        style={{marginTop: '18px'}}
+                        />
+                        <p>0 comments</p>
+                    </Interaction>
+                    <Repost postId={id} token={user.token}/>
             </div>
             <div className="main-post">
                 <div className="top-post">
                     <Link to={`/user/${userId}`}><p><strong>{userInfo.username}</strong></p></Link>
                     <div className="icons">
                         {myPost ? <TiPencil onClick={() => setEditMode(!editMode)} style={{cursor: 'pointer'}}/> : <p></p>}
-                        {isMypost() ? <FiTrash onClick={AbrirModal} style={{marginLeft:'10px'}}/> : <p></p>}
+                        {isMypost() ? <FiTrash onClick={setIsOpen(true)} style={{marginLeft:'10px'}}/> : <p></p>}
                     </div>
                 </div>
                 {editMode ? 
@@ -322,70 +277,3 @@ export default function UserPost(props) {
         </ContainerUserPost>
     )
 }
-
-
-const BoxModal = styled.div`
-    width: 80%;
-    height: 100%;
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-    justify-content: center;
-    div {
-        align-items: center;
-        justify-content: center;
-        flex-direction: row;
-        margin-top: 30px;
-    }
-    button{
-        border-radius: 5px;
-        width: 134px;
-        height: 37px;
-        border: none;
-        font-size: 18px;
-        margin: 10px;
-    }
-`
-
-const ModalTitle = styled.h1`
-font-size: 34px;
-color: #ffffff;
-text-align: center;
-`
-
-const ModalConfirm = styled.button`
-background: #1877F2;
-color: #ffffff;
-opacity: ${props => props.state ? 1 : 0.6};
-cursor: ${props => props.state ? 'pointer' : 'not-allowed'};
-`
-const ModalCancel = styled.button`
-background: #ffffff;
-color: #1877F2;
-opacity: ${props => props.state ? 1 : 0.6};
-cursor: ${props => props.state ? 'pointer' : 'not-allowed'};
-`
-const HashtagCSS = styled.span`
-    font-weight: 700;
-    color: #ffffff;
-`;
-
-const Likes = styled.div`
-    cursor: pointer;  
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-`;
-
-const EditBox = styled.textarea`
-    font-family: 'Lato', sans-serif;
-    font-size: 17px;
-    color: #4c4c4c;
-    margin-bottom: 15px;
-    resize: none;
-    outline: none;
-    border-radius: 5px;
-    padding: 10px;
-    pointer-events: ${props => props.disabled ? 'none' : 'all'};
-    background-color: ${props => props.disabled ? '#e5e5e5' : '#ffffff'};
-`;
